audit: Port patches audit code to a rubocop
